+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this library are documented in this file.
 
 - Introduce `database.sql_helper` as a hacky attempt to ease my ignorance
   with psycopg + sqlalchemy + pandas usage.
+- Introduce `database.with_sqlalchemy_conn` decorator helper.
 - Support decorated `webutil.iemapp` functions that return generators.
 
 ### Bug Fixes